+++ includes/admin/admin-options.php
@@ -2,27 +2,42 @@
 
 if( !defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
 
-// add_action( 'genesis_theme_settings_metaboxes', 'bfg_remove_theme_settings_metaboxes' );
+// add_filter( 'genesis_customizer_theme_settings_config', 'bfg_customizer_theme_settings_config' );
 /**
  * Remove some or all of the options metaboxes in Dashboard > Genesis > Theme Settings.
  *
- * See: http://genesissnippets.com/remove-unused-theme-settings-metaboxes/
- *
  * @since 2.0.0
  */
-function bfg_remove_theme_settings_metaboxes($_genesis_theme_settings_pagehook) {
+function bfg_customizer_theme_settings_config($config) {
 
-	// remove_meta_box( 'genesis-theme-settings-version', $_genesis_theme_settings_pagehook, 'main' );			// Information
-	remove_meta_box( 'genesis-theme-settings-feeds', $_genesis_theme_settings_pagehook, 'main' );				// Custom Feeds
-	remove_meta_box( 'genesis-theme-settings-adsense', $_genesis_theme_settings_pagehook, 'main' ); 			// Google AdSense
-	// remove_meta_box( 'genesis-theme-settings-layout', $_genesis_theme_settings_pagehook, 'main' );			// Default Layout
-	remove_meta_box( 'genesis-theme-settings-header', $_genesis_theme_settings_pagehook, 'main' );				// Header
-	remove_meta_box( 'genesis-theme-settings-nav', $_genesis_theme_settings_pagehook, 'main' );					// Navigation
-	remove_meta_box( 'genesis-theme-settings-breadcrumb', $_genesis_theme_settings_pagehook, 'main' );			// Breadcrumbs
-	// remove_meta_box( 'genesis-theme-settings-comments', $_genesis_theme_settings_pagehook, 'main' );			// Comments and Trackbacks
-	// remove_meta_box( 'genesis-theme-settings-posts', $_genesis_theme_settings_pagehook, 'main' );			// Content Archives
-	// remove_meta_box( 'genesis-theme-settings-blogpage', $_genesis_theme_settings_pagehook, 'main' );			// Blog Page Template
-	// remove_meta_box( 'genesis-theme-settings-scripts', $_genesis_theme_settings_pagehook, 'main' );			// Header and Footer Scripts
+	// Updates
+	unset($config['genesis']['sections']['genesis_updates']);
+
+	// Headers
+	unset($config['genesis']['sections']['genesis_header']);
+
+	// Google AdSense
+	unset($config['genesis']['sections']['genesis_adsense']);
+
+	// Color Scheme
+	unset($config['genesis']['sections']['genesis_color_scheme']);
+
+	// Site Layout
+	unset($config['genesis']['sections']['genesis_layout']);
+
+	// Breadcrumbs
+	unset($config['genesis']['sections']['genesis_breadcrumbs']);
+
+	// Comments and Trackbacks
+	unset($config['genesis']['sections']['genesis_comments']);
+
+	// Content Archives
+	unset($config['genesis']['sections']['genesis_archives']);
+
+	// Header/Footer Scripts
+	unset($config['genesis']['sections']['genesis_scripts']);
+
+	return $config;
 
 }