+++ includes/admin/admin-views.php
@@ -235,7 +235,7 @@ function bfg_user_contactmethods($fields) {
 
 }
 
-// add_action( 'admin_menu', 'bfg_remove_dashboard_menus' );
+add_action( 'admin_menu', 'bfg_remove_dashboard_menus', 12 );
 /**
  * Remove default admin dashboard menus.
  *
@@ -243,17 +243,18 @@ function bfg_user_contactmethods($fields) {
  */
 function bfg_remove_dashboard_menus() {
 
-	remove_menu_page('index.php'); // Dashboard tab
-	remove_menu_page('edit.php'); // Posts
-	remove_menu_page('upload.php'); // Media
-	remove_menu_page('edit.php?post_type=page'); // Pages
-	remove_menu_page('edit-comments.php'); // Comments
-	remove_menu_page('genesis'); // Genesis
-	remove_menu_page('themes.php'); // Appearance
-	remove_menu_page('plugins.php'); // Plugins
-	remove_menu_page('users.php'); // Users
-	remove_menu_page('tools.php'); // Tools
-	remove_menu_page('options-general.php'); // Settings
+	// remove_menu_page('index.php'); // Dashboard tab
+	// remove_menu_page('edit.php'); // Posts
+	// remove_menu_page('upload.php'); // Media
+	// remove_menu_page('edit.php?post_type=page'); // Pages
+	// remove_menu_page('edit-comments.php'); // Comments
+	// remove_menu_page('genesis'); // Genesis
+	remove_submenu_page( 'genesis', 'genesis-plugins' );
+	// remove_menu_page('themes.php'); // Appearance
+	// remove_menu_page('plugins.php'); // Plugins
+	// remove_menu_page('users.php'); // Users
+	// remove_menu_page('tools.php'); // Tools
+	// remove_menu_page('options-general.php'); // Settings
 
 }